notification center, 10.6 support, other fixes
-Updated the base SDK and deployment target so 10.6+ works
--Modified weak references to unsafe_unretained since 10.6 has no
concept of auto-nil weak ref. I haven't gone through to make sure that
the code doesn't do something nasty here now that it doesn't get a free
nil
--Removed auto layout and switched the xibs to springs & struts (auto
layout is not available in 10.6)

-Added notification center support in 10.8 to show current song

-When switching to previous or next song it now resets the song to the
beginning

-If you hit back when a song is more than 5 seconds in it will now go
back to the beginning of that song rather than go the previous track. This
mimics iTunes